Lighting and fmod

+++ engine/src/platform/opengl/gl_mesh.h
@@ -11,7 +11,7 @@ namespace engine
             const std::vector<uint32_t>& indices);
 
         const ref<vertex_array>& va() const override { return m_va; }
-        uint32_t indices_count() const override { return m_indices.size(); }
+        uint32_t indices_count() const override { return (uint32_t)m_indices.size(); }
 
 		std::vector<vertex> vertices() const override { return m_vertices; }
